# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TMS API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Transport Management System"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Permission": {
        "type": "string",
        "enum": [
          "CREATE_USER",
          "READ_USER",
          "UPDATE_USER",
          "DELETE_USER",
          "MANAGE_ROLES",
          "CREATE_TRANSPORT_REQUEST",
          "READ_TRANSPORT_REQUEST",
          "UPDATE_TRANSPORT_REQUEST",
          "DELETE_TRANSPORT_REQUEST",
          "CREATE_MAINTENANCE",
          "READ_MAINTENANCE",
          "UPDATE_MAINTENANCE",
          "DELETE_MAINTENANCE",
          "VIEW_ASSIGNMENTS",
          "CREATE_VEHICLE",
          "READ_VEHICLE",
          "UPDATE_VEHICLE",
          "DELETE_VEHICLE"
        ]
      },
...
      "Area": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "site": {
            "$ref": "#/components/schemas/Site",
            "nullable": true
          },
          "siteId": {
            "type": "string"
          },
          "shafts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Shaft"
            },
            "nullable": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          }
        }
      },
...
      "Exception": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "execution": {
            "$ref": "#/components/schemas/Execution",
            "nullable": true
          },
          "executionId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "evidence": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          }
        }
      },
      "Execution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "procedure": {
            "$ref": "#/components/schemas/Procedure",
            "nullable": true
          },
          "procedureId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User",
            "nullable": true
          },
          "userId": {
            "type": "string"
          },
          "responses": {
            "$ref": "#/components/schemas/JsonValue"
          },
          "status": {
            "type": "string",
            "enum": [
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "startTime": {
            "$ref": "#/components/schemas/Date"
          },
          "exceptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Exception"
            },
            "nullable": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          }
        }
      },
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "siteClassification": {
            "type": "string",
            "enum": [
              "GREEN",
              "ORANGE",
              "RED"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PLANNED",
              "BLOCKED",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobAttachment"
            },
            "nullable": true
          },
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobComment"
            },
            "nullable": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "shaft": {
            "$ref": "#/components/schemas/Shaft",
            "nullable": true
          },
          "shaftId": {
            "type": "string"
          }
        }
      },
      "JobAttachment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "job": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "priority": {
                "type": "string",
                "enum": [
                  "HIGH",
                  "MEDIUM",
                  "LOW"
                ]
              },
              "siteClassification": {
                "type": "string",
                "enum": [
                  "GREEN",
                  "ORANGE",
                  "RED"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "PENDING",
                  "PLANNED",
                  "BLOCKED",
                  "IN_PROGRESS",
                  "COMPLETED",
                  "CANCELLED"
                ]
              },
              "attachments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobAttachment"
                },
                "nullable": true
              },
              "comments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobComment"
                },
                "nullable": true
              },
              "createdAt": {
                "$ref": "#/components/schemas/Date"
              },
              "updatedAt": {
                "$ref": "#/components/schemas/Date"
              },
              "shaft": {
                "$ref": "#/components/schemas/Shaft",
                "nullable": true
              },
              "shaftId": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "jobId": {
            "type": "string"
          },
          "fileUrl": {
            "type": "string"
          },
          "fileType": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          }
        }
      },
      "JobComment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "job": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "priority": {
                "type": "string",
                "enum": [
                  "HIGH",
                  "MEDIUM",
                  "LOW"
                ]
              },
              "siteClassification": {
                "type": "string",
                "enum": [
                  "GREEN",
                  "ORANGE",
                  "RED"
                ]
              },
              "status": {
                "type": "string",
                "enum": [
                  "PENDING",
                  "PLANNED",
                  "BLOCKED",
                  "IN_PROGRESS",
                  "COMPLETED",
                  "CANCELLED"
                ]
              },
              "attachments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobAttachment"
                },
                "nullable": true
              },
              "comments": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JobComment"
                },
                "nullable": true
              },
              "createdAt": {
                "$ref": "#/components/schemas/Date"
              },
              "updatedAt": {
                "$ref": "#/components/schemas/Date"
              },
              "shaft": {
                "$ref": "#/components/schemas/Shaft",
                "nullable": true
              },
              "shaftId": {
                "type": "string"
              }
            },
            "nullable": true
          },
          "jobId": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/User",
            "nullable": true
          },
          "authorId": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          }
        }
      },
      "Maintenance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "vehicle": {
            "$ref": "#/components/schemas/Vehicle",
            "nullable": true
          },
          "vehicleId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "startDate": {
            "$ref": "#/components/schemas/Date"
          },
          "status": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          }
        }
      },
      "Procedure": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "STANDARD",
              "EXCEPTION"
            ]
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "nullable": true
          },
          "executions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Execution"
            },
            "nullable": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          }
        }
      },
...
      "Session": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "$ref": "#/components/schemas/Date"
          },
          "user": {
            "$ref": "#/components/schemas/User",
            "nullable": true
          }
        }
      },
      "Shaft": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "area": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "site": {
                "$ref": "#/components/schemas/Site",
                "nullable": true
              },
              "siteId": {
                "type": "string"
              },
              "shafts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Shaft"
                },
                "nullable": true
              },
              "createdAt": {
                "$ref": "#/components/schemas/Date"
              },
              "updatedAt": {
                "$ref": "#/components/schemas/Date"
              }
            },
            "nullable": true
          },
          "areaId": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "client": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "firstName": {
                "type": "string"
              },
              "lastName": {
                "type": "string"
              },
              "idNumber": {
                "type": "string"
              },
              "address": {
                "type": "string"
              },
              "contactNumber": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "createdAt": {
                "$ref": "#/components/schemas/Date"
              },
              "updatedAt": {
                "$ref": "#/components/schemas/Date"
              },
              "shafts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Shaft"
                },
                "nullable": true
              },
              "syndicates": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Syndicate"
                },
                "nullable": true
              }
            },
            "nullable": true
          },
          "clientId": {
            "type": "string"
          },
          "jobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "nullable": true
          }
        }
      },
      "Syndicate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "clients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Client"
            },
            "nullable": true
          }
        }
      },
      "Site": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "areas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Area"
            },
            "nullable": true
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMINISTRATOR",
              "TRANSPORT_MANAGER",
              "DRIVER",
              "REQUESTER",
              "WORKSHOP_MANAGER"
            ]
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            },
            "nullable": true
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          },
          "assignedJobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "nullable": true
          },
          "requestedJobs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "nullable": true
          },
          "jobComments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobComment"
            },
            "nullable": true
          },
          "executions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Execution"
            },
            "nullable": true
          }
        }
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/VehicleType",
            "nullable": true
          },
          "typeId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "IN_TRANSIT",
              "UNDER_MAINTENANCE",
              "OUT_OF_SERVICE"
            ]
          },
          "maintenanceRecords": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Maintenance"
            },
            "nullable": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          }
        }
      },
      "VehicleType": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "vehicles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehicle"
            },
            "nullable": true
          },
          "createdAt": {
            "$ref": "#/components/schemas/Date"
          },
          "updatedAt": {
            "$ref": "#/components/schemas/Date"
          }
        }
      },
...
      "Date": {
        "type": "string",
        "format": "date-time"
      },
      "JsonValue": {
        "type": "object",
        "additionalProperties": true
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "servers": [
    {
      "url": "https://zfnsz3gffi.execute-api.eu-west-1.amazonaws.com",
      "description": "Development API"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "paths": {
...
    "/user": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "List of users retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get a user by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Update a user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Delete a user",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/vehicle": {
      "post": {
        "tags": [
          "Vehicle"
        ],
        "summary": "Create a new vehicle",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vehicle"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created successfully"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "get": {
        "tags": [
          "Vehicle"
        ],
        "summary": "Get all vehicles",
        "responses": {
          "200": {
            "description": "List of vehicles retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/vehicle/{id}": {
      "get": {
        "tags": [
          "Vehicle"
        ],
        "summary": "Get a vehicle by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vehicle found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "404": {
            "description": "Vehicle not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "tags": [
          "Vehicle"
        ],
        "summary": "Update a vehicle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Vehicle"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vehicle updated successfully"
          },
          "404": {
            "description": "Vehicle not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "tags": [
          "Vehicle"
        ],
        "summary": "Delete a vehicle",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Vehicle deleted successfully"
          },
          "404": {
            "description": "Vehicle not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  },
...
}